refactor(aggregate-sum): mirror PR #663 — split into subdir, reject V0 envelopes
V0 (`MerkOnlyLayerProof`) envelopes predate the aggregate-sum feature and
cannot legitimately carry an aggregate-sum proof, so the V0 layer walker
was unreachable in any honestly-produced proof. Mirror the count-side
changes from PR #663:

  - Convert `aggregate_sum.rs` into a subdirectory mirroring
    `aggregate_count/`: `mod.rs` (public API + V0 rejection),
    `helpers.rs` (envelope decode, single-key layer verification,
    chain enforcement, leaf sum verification), `leaf_chain.rs` (V1
    leaf-chain walker). Removes the dead `verify_v0_layer` path.

  - Add a prover-side V0 gate in `prove_query_non_serialized`: when
    grove_version dispatches to V0 and the path query carries an
    `AggregateSumOnRange` anywhere, return `NotSupported` instead of
    emitting a V0 envelope the verifier would (correctly) reject.

  - Update tests: replace the V0 round-trip test with a V0-rejection
    test; broaden the empty-leaf type-confusion test to accept either
    the V0-rejection or the terminal-type-gate error; remove the now-
    unreachable V0 missing-lower-layer test (V1 counterpart already
    pins the missing-layer behavior); refresh stale doc-comments that
    pointed at `aggregate_sum.rs` line numbers.

No carrier-shape support yet — `AggregateSumOnRange` is still leaf-only
on the merk side, so there is no `classification.rs` / `per_key.rs`
mirror. The leaf-only shape can be extended later in parallel with
matching merk-level work, just like the count side did.
